Allows easy docker testing of UI changes
This is using an "official" docker image that is not the same version we use in production. We only intend to use this for making relatively small UI changes. If we have more extensive work to do, this process should probably not be used.

Add POSTGRES_HOST_AUTH_METHOD to docker file

Co-Authored-By: Jeremy Prevost <JPrevost@users.noreply.github.com>