Fix breadcrumb separator in theme SCSS
Why these changes are being introduced:

The unicode character that we use for breadcrumb separators (`\BB\`) is
not compiled properly during the DSpace 6 build process.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/IN-773

How this addresses that need:

This changes the the separator from the unicode character to `>`.

Side effects of this change:

We don't fully understand the root cause of this problem. It's
possible that similar issues may crop up.