Add Matomo tracking code
Why these changes are being introduced:

We want to track site analytics on Dome.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/UXWS-1568

How this addresses that need:

This adds the Matomo tracking code to the page-structure XSLT.

Side effects of this change:

The site URL and ID are hardcoded, meaning that any future changes
to these values will require an additional commit. We normally prefer
to abstract this information to config variables, but there is not
a particularly intuitive way to do this in DSpace.