Update header structure to use separate links
Rather than have everything in the header link to the platform homepage, we want the Libraries' logo to go to the org homepage, and only the platform name link to the platform home.

This is consistent with how our other application headers work, so this change is correcting a bug that existed previously.